+++ CHANGELOG.md
@@ -14,7 +14,7 @@
   * extensions can now add `container.opts` and files to the docker context
   * Docker template files ending with .fm will are handled as [FreeMarkerTemplate](http://freemarker.org/docs/ref_directives.html)        
   * replace `-fitnesse` option by a plugin again; also removed fitnesse status field
-  
+* dumped logstash extension, it was never used, and we'll have filebeat instead
 
 ### 3.4.10 (pending)